+++ {{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/app/src/client.js
@@ -45,7 +45,8 @@ const render = (appRoutes) => {
                 </ConnectedRouter>
             </Provider>
         </AppContainer>,
-        document.getElementById('root'));
+        document.getElementById('root'),
+    );
 };
 
 if (module.hot) {