+++ {{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/app/src/components/PageError.js
@@ -1,4 +1,4 @@
-import React, {Component} from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import Helmet from 'react-helmet';
 
@@ -7,38 +7,32 @@ import MessagePanel from 'components/MessagePanel';
 import {gettext} from 'utils/i18n';
 
 
-class PageError extends Component {
-    static propTypes = {
-        statusCode: PropTypes.number,
-        title: PropTypes.string,
-        message: PropTypes.string,
-        wide: PropTypes.bool,
-        children: PropTypes.node,
-    };
+const PageError = ({title, message, children, statusCode, wide}) => (
+    <Status code={statusCode}>
+        <div className="container">
+            <MessagePanel title={title} description={message} wide={wide}>
+                <Helmet title={title} />
+                {children}
+            </MessagePanel>
+        </div>
+    </Status>
+);
 
-    static defaultProps = {
-        statusCode: 400,
-        title: gettext('Error'),
-        message: gettext('The server encountered an internal error and was unable to complete your request.'),
-        clear: false,
-        wide: false,
-        children: null,
-    };
+PageError.propTypes = {
+    statusCode: PropTypes.number,
+    title: PropTypes.string,
+    message: PropTypes.string,
+    wide: PropTypes.bool,
+    children: PropTypes.node,
+};
 
-    render() {
-        const {title, message, children, statusCode, wide} = this.props;
-        return (
-            <Status code={statusCode}>
-                <div className="container">
-                    <MessagePanel title={title} description={message} wide={wide}>
-                        <Helmet title={title} />
-                        {children}
-                    </MessagePanel>
-                </div>
-            </Status>
-        );
-    }
-}
+PageError.defaultProps = {
+    statusCode: 400,
+    title: gettext('Error'),
+    message: gettext('The server encountered an internal error and was unable to complete your request.'),
+    wide: false,
+    children: null,
+};
 
 
 export default PageError;