+++ {{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/app/src/server/setup.js
@@ -4,7 +4,7 @@ delete process.env.BROWSER;
 
 global.DEV_MODE = process.env.NODE_ENV !== 'production';
 global.SERVER_MODE = true;
-// Tell `require` calls to look into `/app` also
+// Tell `require` calls to look into correct path in `/app` directory
 process.env.NODE_PATH = DEV_MODE ? 'app/src' : 'app/lib';
 
 require('module').Module._initPaths(); // eslint-disable-line no-underscore-dangle
@@ -13,7 +13,7 @@ require('babel-polyfill');
 require('babel-register')();
 
 // Configure logger early
-require('./logger');
+require('./utils/logger');
 
 // Add fetch to node global context
 global.fetch = require('node-fetch');