+++ {{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/app/src/utils/api.js
@@ -4,7 +4,7 @@ import {getCookie} from 'utils/cookie';
 const resources = Object.entries(DJ_CONST.API || {})
     .filter(item => item[0] !== 'apiRoot')
     .reduce((result, [key, url]) => {
-        result[key] = new Resource(url);  // eslint-disable-line no-param-reassign
+        result[key] = new Resource(url); // eslint-disable-line no-param-reassign
         return result;
     }, {});
 
@@ -18,6 +18,7 @@ const api = new Router(
             'X-CSRFToken': getCookie('csrftoken'),
         }),
         withCredentials: true,
-    });
+    },
+);
 
 export default api;