+++ {{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/app/webpack/config.base.js
@@ -2,9 +2,9 @@
 const fs = require('fs');
 const path = require('path');
 const webpack = require('webpack');
-const ExtractTextPlugin = require('extract-text-webpack-plugin');
+const MiniCssExtractPlugin = require("mini-css-extract-plugin");
 const autoprefixer = require('autoprefixer');
-const StatsPlugin = require('stats-webpack-plugin');
+const StatsWriterPlugin = require("webpack-stats-plugin").StatsWriterPlugin;
 
 const flattenDict = require('./helpers/flatten-dict');
 
@@ -14,20 +14,28 @@ const app_root = path.resolve(__dirname, '..');
 // The django app's dir
 const project_root = path.resolve(app_root, '..');
 
+// Show Deprecation warnings
+process.traceDeprecation = true;
+
 
 function makeConfig(options) {
     const DJ_CONST = JSON.parse(fs.readFileSync(path.join(project_root, 'constants.json')));
 
+    const mode = options.mode || 'none';
     const output = {
         path: path.resolve(app_root, 'build'),
         filename: options.filenameTemplate + '.js',
         chunkFilename: options.filenameTemplate + '.chunk.js',
-        publicPath: DJ_CONST.STATIC_WEBPACK_URL,
+        publicPath: process.env.ENABLE_PROXY === 'y' ? DJ_CONST.STATIC_URL : DJ_CONST.STATIC_WEBPACK_URL,
         library: '{{cookiecutter.repo_name}}',
+        hashFunction: 'sha512',
+        hashDigestLength: 32,
     };
 
     return {
         name: 'client',
+
+        mode,
 
         entry: {
             app: options.prependSources.concat(['babel-polyfill', 'client.js']),
@@ -43,13 +51,16 @@ function makeConfig(options) {
                 use: 'babel-loader',
             }, {
                 test: /\.(css|scss)$/,
-                use: ExtractTextPlugin.extract({
-                    use: [{
-                        loader: "css-loader",
+                use: [
+                    !options.extractCss ? 'style-loader' : MiniCssExtractPlugin.loader,
+                    {
+                        loader: 'css-loader',
                         options: {
-                            sourceMap: true,
-                            minimize: options.minifyCss,
-                        }
+                            options: {
+                                sourceMap: true,
+                                minimize: options.minifyCss,
+                            },
+                        },
                     }, {
                         loader: "postcss-loader",
                         options: {
@@ -65,9 +76,8 @@ function makeConfig(options) {
                             sourceMap: true,
                             outputStyle: 'expanded',
                         }
-                    }],
-                    fallback: 'style-loader',
-                }),
+                    },
+                ],
             }, {
                 test: /\.(jpe?g|png|gif|svg|woff2?|eot|ttf)$/,
                 loader: 'url-loader',
@@ -79,24 +89,9 @@ function makeConfig(options) {
         },
 
         plugins: [
-            new ExtractTextPlugin({
+            new MiniCssExtractPlugin({
                 filename: options.filenameTemplate + '.css',
-                disable: !options.extractCss,
-            }),
-
-            new webpack.optimize.CommonsChunkPlugin({
-                name: 'vendor',
-                minChunks: function (m) {
-                    return /node_modules/.test(m.context);
-                },
-            }),
-            new webpack.optimize.CommonsChunkPlugin({
-                name: 'manifest',
-                // Use hash id for filename to use correct assets
-                filename: process.env.NODE_ENV === 'production' ?
-                    'app/manifest.[hash].js' :
-                    'app/manifest.js',
-                minChunks: Infinity,
+                chunkFilename: options.filenameTemplate + '.chunk.css',
             }),
 
             new webpack.DefinePlugin(flattenDict(DJ_CONST, 'DJ_CONST')),
@@ -104,8 +99,43 @@ function makeConfig(options) {
             new webpack.DefinePlugin({SERVER_MODE: JSON.stringify(false)}),
 
             // Write out stats file to app/build directory, useful for production
-            new StatsPlugin('stats.json'),
+            new StatsWriterPlugin({
+                filename: 'stats.json',
+                fields: [
+                    'errors',
+                    'warnings',
+                    'version',
+                    'hash',
+                    'publicPath',
+                    'outputPath',
+                    'assetsByChunkName',
+                    'modules',
+                ],
+            }),
         ].concat(options.plugins),
+
+        optimization: {
+            namedModules: options.namedModules,
+            minimize: options.minimize,
+
+            splitChunks: {
+                chunks: 'all',
+                name : false,
+
+                cacheGroups: {
+                    default: false,
+                    vendors: {
+                        name: 'vendors',
+                        test: /node_modules/,  // Include all assets in node_modules directory
+                        reuseExistingChunk: true,
+                        enforce: true,
+                        chunks: 'initial',
+                        minChunks: 2,
+                    },
+                },
+            },
+            runtimeChunk: true,
+        },
 
         resolve: {
             modules: ['app/src', 'node_modules'],
@@ -119,7 +149,6 @@ function makeConfig(options) {
 
         devtool: options.devtool,
         target: 'web', // Make web variables accessible to webpack, e.g. window
-        // stats: false, // Don't show stats in the console
 
         performance: options.performance,
     };