+++ {{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/settings/production.py
@@ -7,7 +7,7 @@ ALLOWED_HOSTS = ['{{ cookiecutter.live_hostname }}']
 # Static site url, used when we need absolute url but lack request object, e.g. in email sending.
 SITE_URL = 'https://{{ cookiecutter.live_hostname }}'
 
-KOA_SITE_BASE = SITE_URL
+KOA_DJANGO_BASE = SITE_URL
 DJANGO_SITE_URL = SITE_URL
 
 EMAIL_HOST_PASSWORD = 'TODO (api key)'